requirements.txt
flask==3.0.0
pymysql==1.1.0
pymssql==2.2.7
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
bcrypt==4.0.1
gunicorn==20.1.0
psutil==5.9.5